refactor: 지우개와 펜의 중복 부분 제거
